app/requirements.txt
fastapi
uvicorn
mysql-connector-python
pydantic
python-dotenv
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging 
opentelemetry-instrumentation-mysql 
opentelemetry-exporter-otlp